% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropy.R
\name{getEntropy}
\alias{getEntropy}
\title{Compute entropy of MarkovJump paths at specific times}
\usage{
getEntropy(paths, time)
}
\arguments{
\item{paths}{MarkovJumps paths}

\item{time}{Vector of times at which to evaluate}
}
\description{
Compute entropy of MarkovJump paths at specific times
}
