% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{loadPaths}
\alias{loadPaths}
\title{Loads MarkovJump paths from a CSV file}
\usage{
loadPaths(fileName)
}
\arguments{
\item{fileName}{CSV file name}
}
\value{
MarkovJumps path object
}
\description{
Loads MarkovJump paths from a CSV file
}
