% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{package}
\name{MarkovJumpR}
\alias{MarkovJumpR}
\title{MarkovJumpR: R plotting routines for MarkovJump reconstructions}
\description{
MarkovJumpR is a new toy in the BEAST arsensal to facilitate phylogenetic
reconstruction using travel history information.
}
