% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{orderByWeight}
\alias{orderByWeight}
\title{Create a \code{locationMap} ordered by location weights}
\usage{
orderByWeight(paths, heaviestAtTop = FALSE, minRelativeWeight = 0)
}
\arguments{
\item{paths}{MarkovJump paths for one or more taxa}

\item{heaviestAtTop}{Order locations starting at top of plot}

\item{minRelativeWeight}{Minimum relative weight to include in resulting \code{locationMap}}
}
\value{
\code{locationMap} object
}
\description{
Create a \code{locationMap} ordered by location weights
}
