% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{plotPaths}
\alias{plotPaths}
\title{Plots MarkovJump paths for one taxon}
\usage{
plotPaths(
  paths,
  locationMap = NULL,
  minTime = min(paths$startTime, paths$endTime),
  maxTime = max(paths$startTime, paths$endTime),
  defaultColor = "black",
  otherColor = "black",
  verticalColor = NULL,
  alpha = 0.1,
  yJitterSd = 0.1,
  cex.labels = 0.5,
  addLocationLine = FALSE,
  mustDisplayAllLocations = TRUE,
  xAt = NULL,
  xLabels = TRUE
)
}
\arguments{
\item{paths}{MarkovJumps paths}

\item{locationMap}{A \code{locationMap} object}

\item{minTime}{Minimum time to plot}

\item{maxTime}{Maximum time to plot}

\item{defaultColor}{Default color for dwell times}

\item{otherColor}{Color for locations not provided in \code{locationMap}}

\item{verticalColor}{Color of transitions (default `NULL` uses origin location color)}

\item{alpha}{Alpha value to overlay dwell times}

\item{yJitterSd}{Standard deviation of jitter around location}

\item{cex.labels}{Location text label size}

\item{addLocationLine}{Plot dashed line across locations}

\item{mustDisplayAllLocations}{Check that \code{locationMap} contains all locations in \code{paths}}

\item{xAt}{Time axis \code{at} values}

\item{xLabels}{Time axis \code{labels} values}
}
\value{
MarkovJumps path object
}
\description{
Plots MarkovJump paths for one taxon
}
