% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropy.R
\name{plotEntropy}
\alias{plotEntropy}
\title{Plots entropy for one taxon}
\usage{
plotEntropy(
  entropy,
  time,
  minTime = min(time),
  maxTime = max(time),
  drawBox = FALSE,
  xNoAxis = TRUE,
  xAt = NULL,
  xLabels = TRUE,
  fillColor = NULL,
  lineColor = "black",
  ylab = "Entropy",
  cex.label = 0.75
)
}
\arguments{
\item{entropy}{Vector of entropy values}

\item{time}{Vector of times}

\item{minTime}{Minimum time to plot}

\item{maxTime}{Maximum time to plot}

\item{drawBox}{Add box around plot}

\item{xNoAxis}{Do not show x-axis}

\item{xAt}{Time axis \code{at} values}

\item{xLabels}{Time axis \code{labels} values}
}
\description{
Plots entropy for one taxon
}
